+++ packages/handlersjs-core/lib/handlers/pipe-through.handler.ts
@@ -1,4 +1,4 @@
-import { Observable, of } from 'rxjs';
+import { Observable, of, throwError } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { HandlerArgumentError } from '../errors/handler-argument-error';
 import { Handler } from './handler';
@@ -11,11 +11,7 @@ export class PipeThroughHandler<T, S> extends Handler<T, S> {
 
     super();
 
-    if (!this.handlers) {
-
-      throw new HandlerArgumentError('Argument this.handlers should be set.', this.handlers);
-
-    }
+    if (!handlers) { throw new HandlerArgumentError('Argument handlers should be set.', handlers); }
 
   }
 
@@ -25,7 +21,7 @@ export class PipeThroughHandler<T, S> extends Handler<T, S> {
 
   }
 
-  handle(input: any): Observable<any> {
+  handle(input: any): Observable<S> {
 
     let tempInp = of(input);