+++ packages/handlersjs-core/lib/handlers/sequence.handler.ts
@@ -1,4 +1,4 @@
-import { from, Observable, of } from 'rxjs';
+import { from, Observable, of, throwError } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { HandlerArgumentError } from '../errors/handler-argument-error';
 import { Handler } from './handler';
@@ -9,15 +9,19 @@ export class SequenceHandler<T, S> extends Handler<T, S> {
 
     super();
 
+    if (!handlers) { throw new HandlerArgumentError('Argument handlers should be set.', handlers); }
+
   }
 
-  canHandle(input: T, intermediateOutput: S): Observable<boolean> {
+  canHandle(input: T, intermediateOutput?: S): Observable<boolean> {
 
     return of(true);
 
   }
 
-  handle(input: T, intermediateOutput: S): Observable<S> {
+  handle(input: T, intermediateOutput?: S): Observable<S> {
+
+    intermediateOutput = intermediateOutput ?? { body: null, status: 200, headers: {} } as unknown as S;
 
     return of({ handlers: this.handlers, input, intermediateOutput }).pipe(
       switchMap((data) => from(this.safeHandleMultiple(data.handlers, data.input, data.intermediateOutput))),
@@ -26,12 +30,6 @@ export class SequenceHandler<T, S> extends Handler<T, S> {
   }
 
   private async safeHandleMultiple(handlers: Handler<T, S>[], input: T, intermediateOutput: S): Promise<S> {
-
-    if (!this.handlers) {
-
-      throw new HandlerArgumentError('Argument this.handlers should be set.', this.handlers);
-
-    }
 
     let temporaryIntermediateOutput = intermediateOutput;