+++ packages/handlersjs-core/lib/handlers/waterfall.handler.ts
@@ -1,4 +1,4 @@
-import { from, Observable, of } from 'rxjs';
+import { from, Observable, of, throwError } from 'rxjs';
 import { first, map, switchMap } from 'rxjs/operators';
 import { HandlerArgumentError } from '../errors/handler-argument-error';
 import { Handler } from './handler';
@@ -8,6 +8,8 @@ export class WaterfallHandler<T, S> extends Handler<T, S> {
   constructor(public handlers: Handler<T, S>[]) {
 
     super();
+
+    if (!handlers) { throw new HandlerArgumentError('Argument handlers should be set.', handlers); }
 
   }
 
@@ -19,13 +21,16 @@ export class WaterfallHandler<T, S> extends Handler<T, S> {
 
   handle(input: T, intermediateOutput: S): Observable<S> {
 
+    intermediateOutput = intermediateOutput ?? { body: null, status: 200, headers: {} } as unknown as S;
+
     return of({ input, intermediateOutput, handlers: this.handlers }).pipe(
       switchMap((data) =>
         this.getFirstToHandle(data.input, data.intermediateOutput, data.handlers).pipe(
           map((handlerToExecute) => ({ ...data, handlerToExecute })),
         )),
-      switchMap((data) =>
-        data.handlerToExecute ? data.handlerToExecute.handle(data.input, intermediateOutput) : of(intermediateOutput)),
+      switchMap((data) => data.handlerToExecute
+        ? data.handlerToExecute.handle(data.input, intermediateOutput)
+        : of(intermediateOutput)),
     );
 
   }
@@ -34,19 +39,13 @@ export class WaterfallHandler<T, S> extends Handler<T, S> {
     input: T,
     intermediateOutput: S,
     handlers: Handler<T, S>[],
-  ): Observable<Handler<T, S>> {
-
-    if (!this.handlers) {
-
-      throw new HandlerArgumentError('Argument this.handlers should be set.', this.handlers);
-
-    }
+  ): Observable<Handler<T, S> | undefined> {
 
     return from(handlers).pipe(
       switchMap((handler) =>
         handler.canHandle(input, intermediateOutput).pipe(map((canHandle) => ({ canHandle, handler })))),
       first((result) => result.canHandle, null),
-      map((canHandle) => canHandle?.handler),
+      map((canHandle) => canHandle?.handler)
     );
 
   }