+++ packages/handlersjs-http/lib/servers/node/node-http-request-response.handler.spec.ts
@@ -1,5 +1,5 @@
 import { IncomingMessage, ServerResponse } from 'http';
-import { of } from 'rxjs';
+import { of, throwError } from 'rxjs';
 import * as mockhttp from 'mock-http';
 import { HttpHandler } from '../../models/http-handler';
 import { NodeHttpRequestResponseHandler } from './node-http-request-response.handler';
@@ -12,12 +12,13 @@ describe('NodeHttpRequestResponseHandler', () => {
   let streamMock: NodeHttpStreams;
   let req: IncomingMessage;
   let res: ServerResponse;
+  const buffer = Buffer.from('bodyString');
 
   beforeEach(async () => {
 
     nestedHttpHandler = {
       canHandle: jest.fn(),
-      handle: jest.fn().mockReturnValueOnce(of({ body: {}, headers: {}, status: 200 })),
+      handle: jest.fn().mockReturnValueOnce(of({ body: 'mockBody', headers: { mockKey: 'mockValue' }, status: 200 })),
       safeHandle: jest.fn(),
     } as HttpHandler;
 
@@ -26,6 +27,8 @@ describe('NodeHttpRequestResponseHandler', () => {
     req = new mockhttp.Request({
       url: 'http://localhost:3000/test?works=yes',
       method: 'GET',
+      buffer,
+
     });
 
     res = new mockhttp.Response();
@@ -98,6 +101,22 @@ describe('NodeHttpRequestResponseHandler', () => {
 
     });
 
+    it('should set the body in the request if found in the request stream', async () => {
+
+      await handler.handle(streamMock).toPromise();
+      expect(nestedHttpHandler.handle).toHaveBeenCalledTimes(1);
+
+      expect(nestedHttpHandler.handle).toHaveBeenCalledWith({
+        'request': {
+          'method': 'GET',
+          'headers': {},
+          'url': new URL('http://localhost:3000/test?works=yes'),
+          'body': 'bodyString',
+        },
+      });
+
+    });
+
     it('should call the nested handlers handle method', async () => {
 
       await handler.handle(streamMock).toPromise();
@@ -108,14 +127,14 @@ describe('NodeHttpRequestResponseHandler', () => {
     it('should write the headers to response stream', async () => {
 
       await handler.handle(streamMock).toPromise();
-      expect(streamMock.responseStream.writeHead).toHaveBeenCalledWith(200, {});
+      expect(streamMock.responseStream.writeHead).toHaveBeenCalledWith(200, { mockKey: 'mockValue', 'content-length': Buffer.byteLength('mockBody', 'utf-8').toString() });
 
     });
 
     it('should write the body to response stream', async () => {
 
       await handler.handle(streamMock).toPromise();
-      expect(streamMock.responseStream.write).toHaveBeenCalledWith({});
+      expect(streamMock.responseStream.write).toHaveBeenCalledWith('mockBody');
 
     });
 
@@ -131,9 +150,185 @@ describe('NodeHttpRequestResponseHandler', () => {
       await handler.handle(streamMock).toPromise();
       expect(nestedHttpHandler.handle).toHaveBeenCalledTimes(1);
 
-      expect(nestedHttpHandler.handle).toHaveBeenCalledWith(expect.objectContaining({
-        request: { url: new URL('http://localhost:3000/test?works=yes'), method: 'GET', headers: {} },
-      }));
+      expect(nestedHttpHandler.handle).toHaveBeenCalledWith(expect.objectContaining({ request: expect.objectContaining({ url: new URL('http://localhost:3000/test?works=yes'), method: 'GET', headers: {} }) }));
+
+    });
+
+    it('should catch Errors and create an error response', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(throwError(new Error('mock error')));
+
+      await handler.handle(streamMock).toPromise();
+      const expectedError = 'The server could not process the request due to an error:\nmock error';
+
+      expect(res.writeHead).toHaveBeenCalledWith(500, { 'content-length': Buffer.byteLength(expectedError, 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith(expectedError);
+      expect(res.end).toHaveBeenCalledTimes(1);
+
+    });
+
+    it('should calculate the content-length of the response body as utf-8 when no charset is present', async () => {
+
+      const body = 'This is a response body with a certain length.';
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ body, headers: { 'content-length': '2' }, status:200 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(200, { 'content-length': Buffer.byteLength(body, 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith(body);
+      expect(res.end).toHaveBeenCalledTimes(1);
+
+    });
+
+    it('should calculate the content-length of the response body with the given charset', async () => {
+
+      const body = btoa('This is a response body with a certain length.');
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ body, headers: { 'content-length': '2', 'content-type': 'text/html; charset=base64' }, status:200 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(200, { 'content-length': Buffer.byteLength(body, 'base64').toString(), 'content-type': 'text/html; charset=base64' });
+      expect(res.write).toHaveBeenCalledWith(body);
+      expect(res.end).toHaveBeenCalledTimes(1);
+
+    });
+
+    it('should not calculate content-length when response does not contain a body', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ headers: { mockKey: 'mockValue' }, status:200 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(200, { mockKey: 'mockValue' });
+      expect(res.write).toHaveBeenCalledTimes(0);
+      expect(res.end).toHaveBeenCalledTimes(1);
+
+    });
+
+    it('should error if the charset is not supported', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ body: 'mockBody', headers: { 'content-type': 'text/html; charset=unsupported' }, status:200 }));
+
+      await expect(handler.handle(streamMock).toPromise()).rejects.toThrow('The specified charset is not supported');
+
+    });
+
+    it('should set the charset to utf-8 if no content-type header is specified', async () => {
+
+      const body = 'This is a response body with a certain length.';
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ body, headers: { 'content-type': 'text/html;' }, status:200 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(200, { 'content-length': Buffer.byteLength(body, 'utf-8').toString(), 'content-type': 'text/html;' });
+
+    });
+
+    it('should set the charset to utf-8 if no charset is specified in the content-type header', async () => {
+
+      const body = 'This is a response body with a certain length.';
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ body, headers: { 'content-type': 'text/html;' }, status:200 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(200, { 'content-length': Buffer.byteLength(body, 'utf-8').toString(), 'content-type': 'text/html;' });
+
+    });
+
+    it('should set the charset to utf-8 if no content-type header is specified', async () => {
+
+      const body = 'This is a response body with a certain length.';
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ body, headers: {}, status:200 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(200, { 'content-length': Buffer.byteLength(body, 'utf-8').toString() });
+
+    });
+
+    it('should JSON stringify the response body if the body type is not string', async () => {
+
+      const body = 1234;
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(of({ body, headers: { 'content-type': 'text/html' }, status:200 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(200, { 'content-length': Buffer.byteLength(body.toString(), 'utf-8').toString(), 'content-type': 'text/html' });
+
+    });
+
+    it('should catch objects with a status but without headers and add an empty headers object', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(throwError({ status: 404 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(404, { 'content-length': Buffer.byteLength('Not Found', 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith('Not Found');
+      expect(res.end).toHaveBeenCalledTimes(1);
+
+    });
+
+    it('should catch objects without a status, headers, or message and return an Internal Server Error response', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(throwError({ unknownKey: 'mockValue' }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(500, { 'content-length': Buffer.byteLength('Internal Server Error', 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith('Internal Server Error');
+      expect(res.end).toHaveBeenCalledTimes(1);
+
+    });
+
+    it.each([
+      [ 400, 'Bad Request' ],
+      [ 401, 'Unauthorized' ],
+      [ 403, 'Forbidden' ],
+      [ 404, 'Not Found' ],
+      [ 405, 'Method Not Allowed' ],
+      [ 500, 'Internal Server Error' ],
+    ])('should catch objects with a %i status and header and create a %s error response', async (status, error) => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(throwError({ headers: {}, status }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(status, { 'content-length': Buffer.byteLength(error, 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith(error);
+
+    });
+
+    it('should catch objects with any status that is not matched and if the status is within 400-600 keep the status in the response', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(throwError({ headers: {}, status: 409 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(409, { 'content-length': Buffer.byteLength('An Unexpected Error Occured', 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith('An Unexpected Error Occured');
+
+    });
+
+    it('should catch objects with any status that is not matched and if the status is not within 400-599 return a 500 response', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(throwError({ headers: {}, status: 399 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(500, { 'content-length': Buffer.byteLength('An Unexpected Error Occured', 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith('An Unexpected Error Occured');
+
+    });
+
+    it('should catch objects with any status that is not matched and if the status is not within 400-599 return a 500 response', async () => {
+
+      nestedHttpHandler.handle = jest.fn().mockReturnValueOnce(throwError({ headers: {}, status: 600 }));
+
+      await handler.handle(streamMock).toPromise();
+
+      expect(res.writeHead).toHaveBeenCalledWith(500, { 'content-length': Buffer.byteLength('An Unexpected Error Occured', 'utf-8').toString() });
+      expect(res.write).toHaveBeenCalledWith('An Unexpected Error Occured');
 
     });
 
@@ -176,6 +371,41 @@ describe('NodeHttpRequestResponseHandler', () => {
     it('returns true if input is complete', async () => {
 
       await expect(handler.canHandle(streamMock).toPromise()).resolves.toEqual(true);
+
+    });
+
+  });
+
+  describe('parseBody', () => {
+
+    it('should return the body JSON parsed if content type is application/json', async () => {
+
+      const parsed = (handler as any).parseBody('{"name":"jasper","surname":"vandenberghen"}', 'application/json');
+
+      expect(parsed).toEqual({ name: 'jasper', surname: 'vandenberghen' });
+
+    });
+
+    it('should return the body x-www-form-urlencoded UNPARSED if content type is x-www-form-urlencoded', async () => {
+
+      const parsed = (handler as any).parseBody('name=jasper&surname=vandenberghen', 'application/x-www-form-urlencoded');
+
+      // expect(parsed).toEqual({ name: 'jasper', surname: 'vandenberghen' });
+      expect(parsed).toEqual('name=jasper&surname=vandenberghen');
+
+    });
+
+    it('should return the body if content type is default', async () => {
+
+      const parsed = (handler as any).parseBody('{"name":"jasper","surname":"vandenberghen"}', 'text/plain');
+
+      expect(parsed).toEqual('{"name":"jasper","surname":"vandenberghen"}');
+
+    });
+
+    it('should error when parser fails', () => {
+
+      expect(() => (handler as any).parseBody('{""}', 'application/json')).toThrow(SyntaxError);
 
     });